fix logic [see message for fun description]
Basically: the extension isn't enabled as per the previous check.
So we're checking now if its restricted and if so, does the user
have unlimited power to change it. They don't. So, let's check if
the extension is already enabled. If it is, its good. If its not,
bad. But We don't have it enabled - so surely it being enabled is
bad not good? You'd think so. But I didn't originally so here it is.