Redesign form submission (#281)
Per [T7125](https://phabricator.miraheze.org/T7125)

In addition, this includes huge performance improvements. It now loads ManageWiki almost instantly, supports direct linking to tabs (on load it will go to linked tab, this was broken before), supports no JavaScript with client-nojs styling, fixes a bug with duplicating multiselect widgets, and a few other fixes.

This also adds better error handling. When input validation fails, instead of the form appearing to submit but actually not with no warning to the user except on the tab that the error is contained on, now an error message will appear of the form saying there was errors with some of your input.