Add hook to add sidebar links to managewiki-sidebar-header (#628)
Needed for DataDump to be able to add the sidebar link for it since the order was changed when this was moved to HookHandlers, so the hooks in ManageWiki are loaded after DataDump causing a race condition, and it not adding a link there. It could also have other benefits as well.